refactor: Token 교환 로직 Exchanger로 분리
