+++ src/main/java/com/example/demo/application/TokenExchanger.java
@@ -1,0 +1,8 @@
+package com.example.demo.application;
+
+import com.example.demo.application.dto.OauthToken;
+
+public interface TokenExchanger {
+
+    OauthToken exchange(String authorizationCode);
+}
