+++ src/main/java/com/example/demo/application/dto/OauthToken.java
@@ -1,0 +1,8 @@
+package com.example.demo.application.dto;
+
+public record OauthToken(
+    String idToken,
+    String accessToken,
+    String refreshToken
+) {
+}
