--- src/main/java/com/example/demo/infrastructure/GoogleTokenInfo.java
@@ -1,6 +1,0 @@
-package com.example.demo.infrastructure;
-
-public record GoogleTokenInfo(
-    String id_token
-) {
-}
