+++ src/main/java/com/example/demo/infrastructure/oauth/google/GoogleUserInfo.java
@@ -1,4 +1,4 @@
-package com.example.demo.infrastructure;
+package com.example.demo.infrastructure.oauth.google;
 
 public record GoogleUserInfo(
     String id,
